+++ example/lib/user_model.dart
@@ -16,6 +16,8 @@ class UserModel with _$UserModel {
   factory UserModel.fromJson(Map<String, dynamic> json) =>
       _$UserModelFromJson(json);
 
+  Map<String, dynamic> toJson() => _$UserModelToJson(this);
+
   @ModelField()
   final String id;